refactor: components를 templates로 디렉터리명 변경
- 내부 파일들이 components보다는 DOM template에 가까워 수정함
