+++ src/main.ts
@@ -24,7 +24,7 @@ import {
 } from './ui/renderer.ts';
 import { renderElement, selectElement } from './utils/dom.ts';
 import { RESTAURANTS } from './data/initialData.ts';
-import createRestaurantInfo from './components/RestaurantInfo.ts';
+import createRestaurantInfo from './templates/RestaurantInfo.ts';
 
 addEventListener('load', () => {
   renderHeader();
