+++ src/templates/RestaurantInfo.ts
